refactor(iocuak): move api utils to api folder
